backend/pyproject.toml
[tool.poetry]
name = "helioscta_pjm_da"
version = "0.1.0"
description = "ETL for Helios PJM DA"
authors = ["Your Name <your.email@example.com>"]
packages = [{include = "src"}]
exclude = ["*.logs", "*.png", "*.csv", "*.xlsx", "*.xlsm"]

[tool.poetry.dependencies]
python = ">=3.12,<3.13"
fastapi = ">=0.115.0"
uvicorn = {version = ">=0.34.0", extras = ["standard"]}
psycopg2-binary = "==2.9.10"
python-dotenv = "==1.0.1"
numpy = "==2.0.2"
pandas = ">=2.0.0"
pyarrow = ">=12.0.0"
pytz = "*"
scikit-learn = "*"
tabulate = "*"
plotly = ">=5.0.0"

[tool.pytest.ini_options]
testpaths = ["tests"]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"